more split files and naming
