+++ packages/twenty-server/src/engine/core-modules/upgrade/types/deprecated-since-version.type.ts
@@ -1,7 +1,7 @@
 import {
   ALL_TWENTY_VERSIONS,
-  type AllTwentyVersion,
-} from 'src/engine/core-modules/upgrade/constants/upgrade-command-supported-versions.constant';
+  AllTwentyVersion,
+} from 'src/engine/core-modules/upgrade/constants/twenty-all-versions.constant';
 import { TWENTY_CURRENT_VERSION } from 'src/engine/core-modules/upgrade/constants/twenty-current-version.constant';
 import { IndexOf, IsGreaterOrEqual } from 'twenty-shared/types';
